+++ src/Ext/TechnoType/Hooks.MultiWeapon.cpp
@@ -52,7 +52,7 @@ DEFINE_HOOK(0x7128B2, TechnoTypeClass_ReadINI_MultiWeapon, 0x6)
 
 				for (int weaponIndex : isSecondary)
 				{
-					if (weaponIndex >= weaponCount)
+					if (weaponIndex >= weaponCount || weaponIndex < 0)
 						continue;
 
 					SecondaryList[weaponIndex] = true;