+++ src/New/Entity/LaserTrailClass.cpp
@@ -7,6 +7,9 @@
 // Returns true if drawn, false otherwise.
 bool LaserTrailClass::Update(CoordStruct location)
 {
+	if (Phobos::Config::HideLaserTrailEffects && this->Type->CanBeHidden)
+		return false;
+
 	bool result = false;
 
 	if (!this->LastLocation.isset())