+++ src/New/Type/LaserTrailTypeClass.h
@@ -23,6 +23,7 @@ public:
 	Valueable<bool> CloakVisible;
 	Valueable<bool> CloakVisible_DetectedOnly;
 	Valueable<bool> DroppodOnly;
+	Valueable<bool> CanBeHidden;
 
 	LaserTrailTypeClass(const char* pTitle = NONE_STR) : Enumerable<LaserTrailTypeClass>(pTitle)
 		, DrawType { LaserTrailDrawType::Laser }
@@ -42,6 +43,7 @@ public:
 		, CloakVisible { false }
 		, CloakVisible_DetectedOnly { false }
 		, DroppodOnly { false }
+		, CanBeHidden { true }
 	{ }
 
 	void LoadFromINI(CCINIClass* pINI);